examples/django/requirements.txt
django==1.11
pycloudfn==0.1.115
djangorestframework==3.6.3
